fix(autoops_es): emit os.cgroup.cpu.usage_percent as int64 (long)
process.cpu.percent is always an integer 0-100 (the ES node stats API
returns it as int). os.cgroup.cpu.usage_percent should be consistent so
that both fields stored in the account index have the same type and
precision.

The account index mapping for the field is now `long` (autoops-services
PR #5742 and autoops-bootstrap PR #375). Emitting int64 keeps the stored
value identical to what the agent sends rather than relying on ES type
coercion to truncate a float at ingest time.

Update cache_test.go: change expected type from float64 to int64, swap
InDelta for Equal (exact integer comparison), drop the math import that
was only needed for IsNaN/IsInf guards on the float path. Update
data_test.go: change type assertion from float64 to int64.

+++ x-pack/metricbeat/module/autoops_es/node_stats/cache.go
@@ -109,7 +109,11 @@ func enrichCgroupCpuUsagePercent(node *mapstr.M, prevNode *mapstr.M) {
 	if usageDelta < 0 || periodsDelta <= 0 {
 		return
 	}
-	percent := float64(usageDelta) / (float64(periodsDelta) * float64(quotaMicros) * 1000) * 100
+	// Truncate to a whole-number percentage: process.cpu.percent is already an integer
+	// (the ES node stats API returns 0–100 as int), and the account index mapping for
+	// this field is `long`. Emitting an integer keeps the stored value identical to
+	// what the agent sends rather than relying on ES coercion to truncate a float.
+	percent := int64(float64(usageDelta) / (float64(periodsDelta) * float64(quotaMicros) * 1000) * 100)
 
 	// `setValue` writes a literal flat key; use `Put` so the value lands inside the
 	// nested `os.cgroup.cpu` object that the schema already produces.
